pyproject.toml
[tool.poetry]
name = "django-react-typescript"
description = "This is an non-opinionated Django + React boilerplate built with great development experience and easy deployment in mind."
authors = ["marcelovicentegc <marcelovicentegc@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
sentry-sdk = "^2.5.1"
gunicorn = "^22.0.0"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
cloudinary = "^1.40.0"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
django-better-admin-arrayfield = "^1.4.2"
twilio = "^9.1.1"
psycopg2-binary = "^2.9.9"
pandas = "^3.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
